# applications/services/api/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for school project",
        "title": "Book \u0026 Venyl Loan Service",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/loan/all-by-user/{id}": {
            "get": {
                "description": "Gets a list of all loans by a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Loan"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id of all loans by user",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Loan"
                            }
                        }
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        },
        "/loan/all/": {
            "get": {
                "description": "Gets a list of all loans",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Loan"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Loan"
                            }
                        }
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        },
        "/loan/create/": {
            "post": {
                "description": "Creates a loan entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Loan"
                ],
                "parameters": [
                    {
                        "description": "Create loan",
                        "name": "LoanEntry",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.LoanEntry"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/loan/get/{id}": {
            "get": {
                "description": "Gets a loan by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Loan"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id of loan",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Loan"
                        }
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        },
        "/log/all-by-user/{id}": {
            "get": {
                "description": "Says a list of all logs by user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Log"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id of user",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Log"
                            }
                        }
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        },
        "/log/create/": {
            "post": {
                "description": "Creates a log entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Log"
                ],
                "parameters": [
                    {
                        "description": "Create log entry",
                        "name": "LogEntry",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.LogEntry"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/log/get-by-user/{userId}/{logId}": {
            "get": {
                "description": "Gets a log by user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Log"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id of user",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Id of log",
                        "name": "logId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Log"
                        }
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        },
        "/user/all/": {
            "get": {
                "description": "Gets a list of all users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.User"
                            }
                        }
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        },
        "/user/create/": {
            "post": {
                "description": "Creates a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "parameters": [
                    {
                        "description": "User to create",
                        "name": "User",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/user/get/{id}": {
            "get": {
                "description": "Gets a user by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id of user",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.User"
                        }
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "main.Loan": {
            "type": "object",
            "properties": {
                "entityId": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "main.LoanEntry": {
            "type": "object",
            "properties": {
                "entityId": {
                    "type": "integer"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "main.Log": {
            "type": "object",
            "properties": {
                "entityId": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "unix": {
                    "type": "integer"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "main.LogEntry": {
            "type": "object",
            "properties": {
                "entityId": {
                    "type": "integer"
                },
                "unix": {
                    "type": "integer"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "main.User": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    }
}